add functionality to jvmlanglibrary, decouple Groovy Library from DefaultJavaLibrary - breaks com.facebook.buck.android.AndroidBinaryGraphEnhancerTest, still trying to work out why
